removed edge, renamed translator to painter, minor fixes on application.yml files

+++ artifactory/src/main/java/curly/artifact/controller/ArtifactResourceController.java
@@ -20,24 +20,30 @@ import curly.artifact.model.PagedArtifact;
 import curly.artifact.service.ArtifactService;
 import curly.commons.github.GitHubAuthentication;
 import curly.commons.github.OctoUser;
+import curly.commons.rx.RxResult;
+import curly.commons.web.BadRequestException;
+import curly.commons.web.ModelErrors;
 import lombok.extern.slf4j.Slf4j;
+import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.rest.webmvc.ResourceNotFoundException;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.context.request.async.DeferredResult;
+import rx.Observable;
 
-import java.util.concurrent.Callable;
+import javax.validation.Valid;
 
 import static curly.commons.rx.RxResult.defer;
+import static org.springframework.http.HttpStatus.*;
 import static org.springframework.web.bind.annotation.RequestMethod.*;
 
 /**
@@ -73,6 +79,8 @@ class ArtifactResourceController {
 			produces = MediaType.APPLICATION_JSON_VALUE)
 	public DeferredResult<HttpEntity<Artifact>> artifactResource(@PathVariable("id") String id) {
 		log.trace("Querying single resource based on id {}", id);
+		if (!ObjectId.isValid(id)) throw new BadRequestException("Provided id " + id + " is not valid!");
+
 		return defer(
 				artifactService.findOne(id)
 						.map(o -> o.<ResourceNotFoundException>orElseThrow(ResourceNotFoundException::new))
@@ -81,21 +89,27 @@ class ArtifactResourceController {
 
 
 	@RequestMapping(method = {POST, PUT})
-	public Callable<HttpEntity<?>> saveResource(@RequestBody Artifact artifact,
-												@GitHubAuthentication OctoUser octoUser) {
+	public DeferredResult<HttpEntity<?>> saveResource(@Valid @RequestBody Artifact artifact,
+													  @GitHubAuthentication OctoUser octoUser,
+													  BindingResult bindingResult) {
 		log.debug("Performing save operations based on user {}", octoUser.getId());
+		if (bindingResult.hasErrors()) {
+			return RxResult.defer(Observable.just(new ResponseEntity<>(new ModelErrors(BAD_REQUEST, bindingResult), BAD_REQUEST)));
+		}
 		artifactService.save(artifact, octoUser);
-		return () -> new ResponseEntity<>(HttpStatus.CREATED);
+		return RxResult.defer(Observable.just(new ResponseEntity<>(CREATED)));
 	}
 
 	@RequestMapping(
 			value = "/{id}",
 			method = DELETE)
-	public Callable<HttpEntity<?>> deleteResource(@PathVariable("id") String id,
-												  @GitHubAuthentication OctoUser octoUser) {
+	public DeferredResult<HttpEntity<?>> deleteResource(@PathVariable("id") String id,
+														@GitHubAuthentication OctoUser octoUser) {
 		log.trace("Performing deletion operations based on user {}", octoUser.getId());
+		if (!ObjectId.isValid(id)) throw new BadRequestException("Provided id " + id + " is not valid!");
+
 		artifactService.delete(id, octoUser);
-		return () -> ResponseEntity.noContent().build();
+		return RxResult.defer(Observable.just(new ResponseEntity<>(NO_CONTENT)));
 	}
 
 }
